Open-source '@adeira/flow-config-parser'
Purpose of this commit is to introduce new package with permissive Flow config parser. The motivation is to be able to merge several Flow configs (templates) together into the main one (since Flow doesn't support nested configs similarly to Eslint). So far, I am using only to create a single config version for React Native. However, eventually, I'd like to collect flowconfigs from any project in the monorepo and generate one large flowconfig in the root (so you don't have to specify the things there manually).

It is inspired by this project: https://github.com/gajus/flow-runtime/tree/master/packages/flow-config-parser, however, our parser is up-to-date, more benevolent and works quite differently for purposes of this project (to be able to merge the configs for example).

adeira-source-id: 8b3ab25dd8d264821ed53479588222d3e7e3084a